refactor: rename BaseCard and SnackBar components
